requirements.txt
numpy >= 1.20.1
simpy >= 4.0.1
pyzmq >= 26.0.0
contextlib2 >= 0.6.0
matplotlib >= 3.3.4
pandas >= 2.1.1
openpyxl >= 3.1.5
paho-mqtt
selenium
pywinauto
webdriver_manager
pdfkit
imgkit
types-pytz
mysql>=0.0.3
types-mysqlclient==2.2.0.20250516
pymongo>=4.13.2
redis>=4.5.4
dbutils>=3.0.2
pandas-stubs>=2.3.2
paddlepaddle==3.2.0 -i https://www.paddlepaddle.org.cn/packages/stable/cpu/
paddleocr==3.2.0
pdf2image>=1.17.0
opencv-python-headless
grpcio>=1.76.0
grpcio-tools>=1.76.0